Update bp.py
Added paging code.

+++ bp.py
@@ -7,6 +7,7 @@ parser = argparse.ArgumentParser()
 parser.add_argument("--Token", help="Access Token")
 parser.add_argument("--GH_API_URL", help="Access Token", default="https://api.github.com")
 parser.add_argument("--Organization", help="GitHub Organization")
+parser.add_argument("--Action", help="Set or Delete Branch Protection", default="set")
 parser.add_argument("--Repos", help="Comma separated list. If empty or null, all repos will be considered")
 args = vars(parser.parse_args())
 
@@ -85,8 +86,6 @@ def get_repos():
         if results.status_code != 200:
             raise Exception("Exception Occurred: " + str(results.status_code) + ": " + results.reason + ": " + results.text)  
 
-        #parsed = urlparse.urlparse(results.links.get("last")["url"])
-        #repoCount = int(parse_qs(parsed.query)["page"][0])
         for result in results.json() :
             if result["name"] != ".github":
                 gh_repos.append(result["full_name"])
@@ -94,25 +93,67 @@ def get_repos():
     return gh_repos
     
     
-def main():
+def getPageCount(itemCount, limit):
+    pageCount = 0
+    if itemCount < limit:
+        pageCount = 1
+    else:
+        rem = itemCount % limit
+        if rem > 0:
+            pageCount = (itemCount // limit) + 1
+        else:
+            pageCount = (itemCount // limit)
+    
+    return pageCount      
+
+
+def getAllRepos():
+    limit = 100
+    gh_repos = []
+    if args["Repos"] != None:
+        gh_repos = args["Repos"].split(',')
+    else:
+        url = args["GH_API_URL"] + "/orgs/" + args["Organization"] + "/repos?type=all&per_page=1"
+        repoCntResult = getRepoResults(url)
+        parsed = urlparse.urlparse(repoCntResult.links.get("last")["url"])
+        repoCount = int(parse_qs(parsed.query)["page"][0])
+        pageCount = getPageCount(repoCount, limit)
+        for pageNum in range(pageCount):
+            pageNum = pageNum + 1
+            url = args["GH_API_URL"] + "/orgs/" + args["Organization"] + "/repos?type=all&per_page=" + str(limit) + "&page=" + str(pageNum)
+            results = getRepoResults(url)
+            for result in results.json():
+                if result["name"] != ".github":
+                    gh_repos.append(result["full_name"])
+
+    return gh_repos     
+
+def getRepoResults(url):
+    results = requests.get(url, headers=make_api_headers())
+    if results.status_code != 200:
+        raise Exception("Exception Occurred: " + str(results.status_code) + ": " + results.reason + ": " + results.text)  
+
+    return results
+
+def main():    
     print("Entering Main function inside Python")
     try:
-        print(args["GH_API_URL"])
-        repos = get_repos()
+        print("Action: " + args["Action"])
+        repos = getAllRepos()
         f = open('bp.json',)
         data = json.load(f)
         for repo in repos:
-            print(repo)
+            print("Repo: " + repo)
             branches = get_branches(repo)
             for branch in branches:
-                print(branch)               
-                #get_protection(repo, branch)
-                #print("Setting Branch Protection")
-                #set_protection(repo, branch, data)                
-                #print("Set Branch Protection Succesfully.")
-                print("Deleting Branch Protection")
-                delete_protection(repo, branch)
-                print("Deleted Branch Protection Succesfully.")
+                if args["Action"] == "set":
+                    print("Setting Branch Protection for " + branch)
+                    set_protection(repo, branch, data)
+                    print("Set Branch Protection Succesfully for " + branch)
+                elif args["Action"] == "delete":
+                    print("Deleting Branch Protection for " + branch)
+                    delete_protection(repo, branch)
+                    print("Deleted Branch Protection Succesfully for " + branch)                
     
     except:
         raise
